requirements.txt
fastapi~=0.135.0
uvicorn[standard]~=0.38.0
gunicorn~=23.0.0
convertdate~=2.4.1
pydantic>=2.0
pytest~=9.0.2
httpx
sqlalchemy>=2.0
psycopg[binary]
python-multipart
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
